Add `--ignore-case` command line option
