+++ main.go
@@ -20,6 +20,7 @@ var debug bool
 var fieldSeparator string = " "
 var files []string = []string{}
 var forceOutput bool
+var ignoreCase bool
 var ignoreLeadingBlanks bool
 var key KeyType
 var memprofile string
@@ -62,6 +63,7 @@ F1,F2  Use all fields between [F1,F2] for comparison
 	flag.BoolVar(&debug, "debug", false, "Print debugging output")
 	flag.StringVar(&fieldSeparator, "field-separator", " ", "Use this field separator")
 	flag.BoolVar(&forceOutput, "force", false, "Overwrite output file if it exists")
+	flag.BoolVar(&ignoreCase, "ignore-case", false, "Ignore case for comparisons")
 	flag.BoolVar(&ignoreLeadingBlanks, "ignore-leading-blanks", false, "Ignore leading whitespace")
 	flag.BoolVar(&ignoreLeadingBlanks, "ignore-leading-whitespace", false, "Ignore leading whitespace, same as --ignore-leading-blanks")
 	flag.Var(&key, "key", "Sort lines based on a particular field, see 'Key Specification' for details")
