+++ process_test.go
@@ -29,7 +29,9 @@ func TestProcessInputFiles1(t *testing.T) {
 		{Lines: []string{"Line four"}, Fields: []string{"Line", "four"}, CompareLine: "Line four"},
 	}
 	var got ContentType = ProcessInputFiles(input, key)
-	compareContentTypes(got, expected, t)
+	if !reflect.DeepEqual(got, expected) {
+		t.Errorf("got\n%s\nexpected\n%s", got, expected)
+	}
 }
 
 func TestProcessInputFiles2(t *testing.T) {
@@ -48,7 +50,9 @@ func TestProcessInputFiles2(t *testing.T) {
 		{Lines: []string{"Line three", "Line four"}, Fields: []string{"Line", "three", "Line", "four"}, CompareLine: "Line three Line four"},
 	}
 	var got ContentType = ProcessInputFiles(input, key)
-	compareContentTypes(got, expected, t)
+	if !reflect.DeepEqual(got, expected) {
+		t.Errorf("got\n%s\nexpected\n%s", got, expected)
+	}
 }
 
 func TestProcessInputFiles3(t *testing.T) {
@@ -68,25 +72,31 @@ func TestProcessInputFiles3(t *testing.T) {
 		{Lines: []string{" Line three", "Line four"}, Fields: []string{"Line", "three", "Line", "four"}, CompareLine: "Line"},
 	}
 	var got ContentType = ProcessInputFiles(input, key)
-	compareContentTypes(got, expected, t)
+	if !reflect.DeepEqual(got, expected) {
+		t.Errorf("got\n%s\nexpected\n%s", got, expected)
+	}
 }
 
-// func TestProcessInputFiles3(t *testing.T) {
-	// var input []string = []string{
-		// "Line 1 a",
-		// "Line 2 b",
-		// "Line 3 c",
-		// "Line 4 d",
-	// }
-	// multiline = 3
-	// var key KeyType = KeyType{
-		// Type: SingleField,
-		// Keys: []int{2},
-	// }
-	// var expected ContentType = ContentType{
-		// {Lines: []string{"Line one", "Line two"}, Fields: []string{"Line", "one", "Line", "two"}, CompareLine: "Line one Line two"},
-		// {Lines: []string{"Line three", "Line four"}, Fields: []string{"Line", "three", "Line", "four"}, CompareLine: "Line three Line four"},
-	// }
-	// var got ContentType = ProcessInputFiles(input, key)
-	// compareContentTypes(got, expected, t)
-// }
+func TestProcessInputFiles4(t *testing.T) {
+	var input []string = []string{
+		"line 1",
+		"Line 2",
+		"line 3",
+		"Line 4",
+		"line 5",
+	}
+	var expectedCase ContentType = ContentType{
+		{Lines: []string{"line 1"}, Fields: []string{"line", "1"}, CompareLine: "line 1"},
+		{Lines: []string{"Line 2"}, Fields: []string{"Line", "2"}, CompareLine: "line 2"},
+		{Lines: []string{"line 3"}, Fields: []string{"line", "3"}, CompareLine: "line 3"},
+		{Lines: []string{"Line 4"}, Fields: []string{"Line", "4"}, CompareLine: "line 4"},
+		{Lines: []string{"line 5"}, Fields: []string{"line", "5"}, CompareLine: "line 5"},
+	}
+	var key = KeyType{}
+	multiline = 1
+	ignoreCase = true
+	var got ContentType = ProcessInputFiles(input, key)
+	if !reflect.DeepEqual(got, expectedCase) {
+		t.Errorf("got\n%s\nexpected\n%s", got, expectedCase)
+	}
+}
